+++ apps/admin/src/app/(auth)/layout.tsx
@@ -1,8 +1,8 @@
 'use client';
 
+import React from 'react';
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
-import React from 'react';
 
 export default function RootLayout({
   children,
