+++ apps/admin/src/app/(root)/dashboard/[_id]/layout.tsx
@@ -1,13 +1,12 @@
 'use client';
 
-import { Button } from '@admin/components/ui/button';
-import { signOut } from 'next-auth/react';
+import { Navbar } from '@admin/components';
 import React from 'react';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <React.Fragment>
-      <Button onClick={() => signOut()}>Salir</Button>
+      <Navbar />
       {children}
     </React.Fragment>
   );
