+++ apps/admin/src/components/index.tsx
@@ -1,0 +1,4 @@
+export * from './forms';
+export * from './modals';
+export * from './shared';
+export * from './ui';
