+++ apps/admin/src/components/shared/Navbar.tsx
@@ -1,0 +1,3 @@
+export function Navbar() {
+  return <div>This will be a navbar</div>;
+}
